изменил наименование БД + подкорректировал обработку в checkin через базу

+++ auth/checkin.php
@@ -2,13 +2,12 @@
 require_once('../Cors.php');
 $cors = new Cors();
 $cors->cors_policy();
-
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
   try {
     require_once 'db.php';
     $config = parse_ini_file('../config.ini', true);
-    $db = new DB_Conncetion($config['DB']);
-    $db->db_connect();
+    $db = new DB_Connection($config['DB']);
+    $db_connect = $db->db_connect();
     $username = $_POST['username'];
     $password = $_POST['password'];
     $user_role = $_POST['role'];
@@ -21,38 +20,34 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
       exit();
     }
     $parsed_result = pg_fetch_assoc($result);
+    $result_data = array();
     if (pg_num_rows($result) > 0) {
       // добавление в таблицу пользователей
-      $req_insert = 'INSERT INTO accounts.users (username, password_hash, created_at, role)
-            VALUES ($1, $2, current_timestamp, $3)
-            ON CONFLICT(role)
-            DO UPDATE SET
-            username = EXCLUDED.username,
-            password_hash = EXCLUDED.password_hash,
-            created_at = EXCLUDED.created_at';
-
-      $stmt =  pg_prepare($db->get_conn(), "upsert", $req_insert);
-      $upsert_result = pg_execute($db->get_conn(), "upsert", array($username, $hashed_password, $user_role));
-
+      $stmt = "select accounts.f_insert_new_users ('" . $username . "','" . $hashed_password  . "','" . $user_role  . "')";
+      $query = pg_query($db_connect, $stmt);
+      if (!$query) {
+        throw new Exception('Error during DB query insert into temp_users');
+      }
+      $result = pg_fetch_array($query);
+      array_push($result_data,  array("status" => "ok", "inserted_on" => $result[0]));
       // удаляем из временной таблицы
       $req_delete = 'DELETE FROM accounts.temp_users
-        WHERE role=$1;
+        WHERE username=$1;
       ';
       // удаление
       $stmt_del =  pg_prepare($db->get_conn(), "delete", $req_delete);
       pg_execute(
         $db->get_conn(),
         "delete",
-        array($user_role)
+        array($username)
       );
+      echo json_encode($result_data, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
       http_response_code(200);
-      echo json_encode(['message' => 'Successfull']);
-      session_write_close();
-      exit();
     } else {
       http_response_code(403);
       echo json_encode(['Error' => 'There are no users!']);
       session_write_close();
+      $db->closeConn();
       exit();
     }
   } catch (Exception $e) {
